requirements.txt
fastapi
pydantic
requests
openai
python-dotenv
uvicorn